requirements.txt
alembic==0.8.10
appdirs==1.4.0
coverage==4.0.3
decorator==4.0.11
Flask==0.10.1
Flask-Login==0.4.0
Flask-Migrate==1.7.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gunicorn==19.7.1
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.9.5
linecache2==1.0.0
lxml==3.4.0
Mako==1.0.6
MarkupSafe==0.23
nltk==3.2.2
packaging==16.8
pbr==1.10.0
phonenumbers==7.2.7
pyparsing==2.1.10
python-editor==1.0.3
pytz==2016.10
six==1.10.0
SQLAlchemy==1.0.11
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
textblob==0.12.0
traceback2==1.4.0
twilio==5.3.0
unittest2==1.1.0
Werkzeug==0.11.15
WTForms==2.1
xmlunittest==0.3.2
